requirements.txt
ultralytics
opencv-python
requests
pillow
qrcode[pil]
paho-mqtt
pyserial
customtkinter